feat: 마이페이지의 api, types가 독립된 도메인을 갖도록 수정
