+++ apps/web/types/my.ts
@@ -1,0 +1,5 @@
+import { Vote } from "./vote";
+
+export interface MyPageVote extends Vote {
+  countComment: number;
+}
